:pushpin: Fix: 공통 에러 & 응답 처리 수정
