+++ src/main/java/what/what2eat/global/exception/GlobalExceptionHandler.java
@@ -8,26 +8,17 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.NoHandlerFoundException;
 import what.what2eat.global.response.ApiResponse;
-import what.what2eat.global.response.ErrorCode;
 
 @Slf4j
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
-    // 존재하지 않는 요청에 대한 예외
-    @ExceptionHandler(value = {NoHandlerFoundException.class, HttpRequestMethodNotSupportedException.class})
-    public ResponseEntity<ApiResponse<String>> handleNoPageFoundException(Exception e) {
-        log.error("GlobalExceptionHandler catch NoHandlerFoundException : {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(ApiResponse.fail(new CustomException(ErrorCode.NOT_FOUND)));
-    }
-
     // 커스텀 예외
     @ExceptionHandler(value = {CustomException.class})
     public ResponseEntity<ApiResponse<String>> handleCustomException(CustomException e) {
         log.error("handleCustomException() in GlobalExceptionHandler throw CustomException : {}", e.getMessage());
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.fail(new CustomException(ErrorCode.BAD_REQUEST)));
+        return ResponseEntity.status(e.getErrorCode().getHttpStatus())
+                .body(ApiResponse.fail(new CustomException(e.getErrorCode())));
     }
 
     // 기본 예외
